+++ R/explore.R
@@ -50,9 +50,10 @@
 #'  Defaults to 2.
 #' @param override A vector of signals for which the user intends to manually
 #'  define which movement events are valid and invalid.
-#' @param plot.detections.by The type of y axis desired for the individual
+#' @param detections.y.axis The type of y axis desired for the individual
 #'  detection plots. While the argument defaults to "auto", it can be hard-set 
 #'  to one of "stations" or "arrays".
+#' @param plot.detections.by DEPRECATED. Please use the argument detections.y.axis instead.
 #' @param print.releases Logical: Should the release sites be printed in the
 #'  study area diagrams?
 #' @param report Logical. Should an HTML report be created at the end of the
@@ -152,8 +153,13 @@ explore <- function(
   GUI = c("needed", "always", "never"),
   save.tables.locally = FALSE,
   print.releases = TRUE,
-  plot.detections.by = c("auto", "stations", "arrays")) 
+  plot.detections.by,
+  detections.y.axis = c("auto", "stations", "arrays")) 
 {
+
+# check deprecated argument
+  if (!missing(plot.detections.by))
+    stop("'plot.detections.by' has been deprecated. Please use 'detections.y.axis' instead.", call. = FALSE)
 
 # clean up any lost helpers
   deleteHelpers()
@@ -194,14 +200,14 @@ explore <- function(
                         exclude.tags = exclude.tags,
                         override = override,
                         print.releases = print.releases,
-                        plot.detections.by = plot.detections.by)
+                        detections.y.axis = detections.y.axis)
   
   speed.method <- aux$speed.method
   speed.warning <- aux$speed.warning
   speed.error <- aux$speed.error
   inactive.warning <- aux$inactive.warning
   inactive.error <- aux$inactive.error
-  plot.detections.by <- aux$plot.detections.by
+  detections.y.axis <- aux$detections.y.axis
   rm(aux)
 
   GUI <- checkGUI(GUI, save.tables.locally = save.tables.locally)
@@ -231,7 +237,7 @@ explore <- function(
     ", GUI = '", GUI, "'",
     ", save.tables.locally = '", ifelse(save.tables.locally, "TRUE", "FALSE"),
     ", print.releases = ", ifelse(print.releases, "TRUE", "FALSE"),
-    ", plot.detections.by = '", plot.detections.by, "'",
+    ", detections.y.axis = '", detections.y.axis, "'",
     ")")
 
   appendTo("debug", the.function.call)
@@ -465,7 +471,7 @@ explore <- function(
                                          valid.movements = valid.movements,
                                          spatial = spatial,
                                          rsp.info = rsp.info,
-                                         type = plot.detections.by)
+                                         y.axis = detections.y.axis)
 
     circular.plots <- printCircular(times = timesToCircular(times), 
                                     bio = bio)
@@ -474,7 +480,7 @@ explore <- function(
       sensor.plots <- printSensorData(detections = valid.detections, 
                                       spatial = spatial,
                                       rsp.info = rsp.info, 
-                                      type = plot.detections.by)
+                                      colour.by = detections.y.axis)
     } else {
       sensor.plots <- NULL
     }
@@ -525,7 +531,7 @@ explore <- function(
                     deployments = deployments,
                     detections = detections,
                     valid.detections = valid.detections,
-                    plot.detections.by = plot.detections.by)
+                    detections.y.axis = detections.y.axis)
 
     appendTo("debug", "debug: Converting report to html")
     rmarkdown::render(input = paste0(tempdir(), "/actel_report_auxiliary_files/actel_explore_report.Rmd"),
@@ -594,7 +600,7 @@ explore <- function(
 #' @keywords internal
 #'
 printExploreRmd <- function(override.fragment, biometric.fragment, individual.plots,
-  circular.plots, sensor.plots, spatial, deployments, detections, valid.detections, plot.detections.by){
+  circular.plots, sensor.plots, spatial, deployments, detections, valid.detections, detections.y.axis){
 
   work.path <- paste0(tempdir(), "/actel_report_auxiliary_files/")
 
@@ -607,7 +613,7 @@ printExploreRmd <- function(override.fragment, biometric.fragment, individual.pl
     sensor.fragment <- paste0("### Sensor plots
 
 Note:
-  : The colouring in these plots will follow that of the individual detection plots, which can be modified using `plot.detections.by`.
+  : The colouring in these plots will follow that of the individual detection plots, which can be modified using `detections.y.axis`.
   : The data used for these graphics is stored in the `valid.detections` object.
   : You can replicate these graphics and edit them as needed using the `plotSensors()` function.
 
@@ -724,11 +730,11 @@ Note:
 ### Individual detection plots
 
 Note:
-  : You can choose to plot detections by station or by array using the `plot.detections.by` argument.
-  : The detections are coloured by ', ifelse(plot.detections.by == "stations", 'array', 'section'), '. The vertical black dashed line shows the release time. The full dark-grey line shows the movement events considered valid, while the dashed dark-grey line shows the movement events considered invalid.
-', ifelse(plot.detections.by == "stations", '  : The movement event lines move straight between the first and last station of each event (i.e. in-between detections will not be individually linked by the line).\n', ''),
+  : You can choose to plot detections by station or by array using the `detections.y.axis` argument.
+  : The detections are coloured by ', ifelse(detections.y.axis == "stations", 'array', 'section'), '. The vertical black dashed line shows the release time. The full dark-grey line shows the movement events considered valid, while the dashed dark-grey line shows the movement events considered invalid.
+', ifelse(detections.y.axis == "stations", '  : The movement event lines move straight between the first and last station of each event (i.e. in-between detections will not be individually linked by the line).\n', ''),
 '  : Manually **edited** tags are highlighted with **yellow** graphic borders.
-  : The ', ifelse(plot.detections.by == "stations", 'stations', 'arrays'), ' have been aligned by ', ifelse(plot.detections.by == "stations", 'array', 'section'), ', following the order provided ', ifelse(plot.detections.by == "stations", '', 'either '), 'in the spatial input', ifelse(plot.detections.by == "stations", '.', ' or the `section.order` argument.'), '
+  : The ', ifelse(detections.y.axis == "stations", 'stations', 'arrays'), ' have been aligned by ', ifelse(detections.y.axis == "stations", 'array', 'section'), ', following the order provided ', ifelse(detections.y.axis == "stations", '', 'either '), 'in the spatial input', ifelse(detections.y.axis == "stations", '.', ' or the `section.order` argument.'), '
   : You can replicate these graphics and edit them as needed using the `plotDetections()` function.
   : You can also see the movement events of multiple tags simultaneously using the `plotMoves()` function.
   : The data used in these graphics is stored in the `detections` and `movements` objects (and respective valid counterparts).