+++ R/migration.R
@@ -147,8 +147,13 @@ migration <- function(
   GUI = c("needed", "always", "never"),
   save.tables.locally = FALSE,
   print.releases = TRUE,
-  plot.detections.by = c("auto", "stations", "arrays"))
+  plot.detections.by,
+  detections.y.axis = c("auto", "stations", "arrays"))
 {
+
+# check deprecated argument
+  if (!missing(plot.detections.by))
+    stop("'plot.detections.by' has been deprecated. Please use 'detections.y.axis' instead.", call. = FALSE)
 
 # clean up any lost helpers
   deleteHelpers()
@@ -193,14 +198,14 @@ migration <- function(
                         if.last.skip.section = if.last.skip.section,
                         replicates = replicates,
                         section.order = section.order,
-                        plot.detections.by = plot.detections.by)
+                        detections.y.axis = detections.y.axis)
 
   speed.method <- aux$speed.method
   speed.warning <- aux$speed.warning
   speed.error <- aux$speed.error
   inactive.warning <- aux$inactive.warning
   inactive.error <- aux$inactive.error
-  plot.detections.by <- aux$plot.detections.by
+  detections.y.axis <- aux$detections.y.axis
   rm(aux)
 
   GUI <- checkGUI(GUI, save.tables.locally)
@@ -235,7 +240,7 @@ migration <- function(
     ", GUI = '", GUI, "'",
     ", save.tables.locally = ", ifelse(save.tables.locally, "TRUE", "FALSE"),
     ", print.releases = ", ifelse(print.releases, "TRUE", "FALSE"),
-    ", plot.detections.by = '", plot.detections.by, "'",
+    ", detections.y.axis = '", detections.y.axis, "'",
     ")")
 
   appendTo("debug", the.function.call)
@@ -664,7 +669,7 @@ migration <- function(
                                          spatial = spatial,
                                          status.df = status.df,
                                          rsp.info = rsp.info,
-                                         type = plot.detections.by)
+                                         y.axis = detections.y.axis)
 
     circular.plots <- printCircular(times = timesToCircular(times),
                                     bio = bio)
@@ -676,7 +681,7 @@ migration <- function(
       sensor.plots <- printSensorData(detections = valid.detections, 
                                       spatial = spatial,
                                       rsp.info = rsp.info, 
-                                      type = plot.detections.by)
+                                      colour.by = detections.y.axis)
     } else {
       sensor.plots <- NULL
     }
@@ -732,7 +737,7 @@ migration <- function(
                         deployments = deployments,
                         valid.detections = valid.detections,
                         detections = detections,
-                        plot.detections.by = plot.detections.by)
+                        detections.y.axis = detections.y.axis)
 
     appendTo("debug", "debug: Converting report to html")
     rmarkdown::render(input = paste0(tempdir(), "/actel_report_auxiliary_files/actel_migration_report.Rmd"),
@@ -813,7 +818,7 @@ migration <- function(
 printMigrationRmd <- function(override.fragment, biometric.fragment, section.overview,
   efficiency.fragment, display.progression, array.overview.fragment, survival.graph.size,
   individual.plots, circular.plots, sensor.plots, spatial, deployments, valid.detections, 
-  detections, plot.detections.by){
+  detections, detections.y.axis){
 
   work.path <- paste0(tempdir(), "/actel_report_auxiliary_files/")
 
@@ -826,7 +831,7 @@ printMigrationRmd <- function(override.fragment, biometric.fragment, section.ove
     sensor.fragment <- paste0("### Sensor plots
 
 Note:
-  : The colouring in these plots will follow that of the individual detection plots, which can be modified using `plot.detections.by`.
+  : The colouring in these plots will follow that of the individual detection plots, which can be modified using `detections.y.axis`.
   : The data used for these graphics is stored in the `valid.detections` object.
   : You can replicate these graphics and edit them as needed using the `plotSensors()` function.
 
@@ -995,12 +1000,12 @@ Note:
 ### Individual detection plots
 
 Note:
-  : You can choose to plot detections by station or by array using the `plot.detections.by` argument.
-  : The detections are coloured by ', ifelse(plot.detections.by == "stations", 'array', 'section'), '. The vertical black dashed line shows the time of release. The vertical grey dashed lines show the assigned moments of entry and exit for each study area section. The full dark-grey line shows the movement events considered valid, while the dashed dark-grey line shows the movement events considered invalid.
-', ifelse(plot.detections.by == "stations", '  : The movement event lines move straight between the first and last station of each event (i.e. in-between detections will not be individually linked by the line).\n', ''),
+  : You can choose to plot detections by station or by array using the `detections.y.axis` argument.
+  : The detections are coloured by ', ifelse(detections.y.axis == "stations", 'array', 'section'), '. The vertical black dashed line shows the time of release. The vertical grey dashed lines show the assigned moments of entry and exit for each study area section. The full dark-grey line shows the movement events considered valid, while the dashed dark-grey line shows the movement events considered invalid.
+', ifelse(detections.y.axis == "stations", '  : The movement event lines move straight between the first and last station of each event (i.e. in-between detections will not be individually linked by the line).\n', ''),
 '  : Manually **edited** tag detections are highlighted with **yellow** graphic borders.
   : Manually **overridden** tag detections are highlighted with **red** graphic borders.
-  : The ', ifelse(plot.detections.by == "stations", 'stations', 'arrays'), ' have been aligned by ', ifelse(plot.detections.by == "stations", 'array', 'section'), ', following the order provided ', ifelse(plot.detections.by == "stations", '', 'either '), 'in the spatial input', ifelse(plot.detections.by == "stations", '.', ' or the `section.order` argument.'), '
+  : The ', ifelse(detections.y.axis == "stations", 'stations', 'arrays'), ' have been aligned by ', ifelse(detections.y.axis == "stations", 'array', 'section'), ', following the order provided ', ifelse(detections.y.axis == "stations", '', 'either '), 'in the spatial input', ifelse(detections.y.axis == "stations", '.', ' or the `section.order` argument.'), '
   : You can replicate these graphics and edit them as needed using the `plotDetections()` function.
   : You can also see the movement events of multiple tags simultaneously using the `plotMoves()` function.
   : The data used in these graphics is stored in the `detections` and `movements` objects (and respective valid counterparts).